08_board user N:1 model relation

+++ boards/views.py
@@ -34,7 +34,10 @@ def create(request):
     if request.method == 'POST':
         form = BoardForm(request.POST)
         if form.is_valid():
-            board = form.save()
+            board = form.save(commit=False) # board 를 바로 저장하지 않고, 현재 user 를 넣고 저장
+            #실제 DB 반영 전까지의 단계를 진행하고, 그 중간에 user 정보를 requser.user 에서 가져와서 그 후에 저장한다.
+            board.user = request.user
+            board.save()
             return redirect('boards:detail', board.pk)
     else:
         form = BoardForm()
@@ -53,11 +56,14 @@ def detail(request, board_pk):
     
 def delete(request, board_pk):
     board = get_object_or_404(Board, pk=board_pk)
-    if request.method == 'POST':
-        board.delete()
-        return redirect('boards:index')
+    if board.user == request.user:
+        if request.method == 'POST':
+            board.delete()
+            return redirect('boards:index')
+        else:
+            return redirect('boards:detail', board.pk)
     else:
-        return redirect('boards:detail', board.pk)
+        return redirect('boards:index')
         
         
         
@@ -85,14 +91,17 @@ def delete(request, board_pk):
 @login_required
 def update(request, board_pk):
     board = get_object_or_404(Board, pk=board_pk)
-    if request.method == 'POST':
-        form = BoardForm(request.POST, instance=board)  # 1
-        if form.is_valid():
-            board = form.save()                         # 2
-            return redirect('boards:detail', board.pk)
-            
+    if board.user == request.user:
+        if request.method == 'POST':
+            form = BoardForm(request.POST, instance=board)  # 1
+            if form.is_valid():
+                board = form.save()                         # 2
+                return redirect('boards:detail', board.pk)
+                
+        else:
+            form = BoardForm(instance=board)                 # 3
     else:
-        form = BoardForm(instance=board)                 # 3
+        return redirect('boards:index')
     context = {
         'form' : form,
         'board' : board,
